+++ packages/xsoulspace_foundation/lib/src/foundation.dart
@@ -1,0 +1,10 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+export 'annotations.dart';
+export 'extensions/extensions.dart';
+export 'hooks/hooks.dart';
+export 'interfaces.dart';
+export 'utils/utils.dart';
+
+part 'foundation.freezed.dart';
+part 'models/models.dart';